+++ paver/path25.py
@@ -11,8 +11,20 @@ __all__ = ['path', 'pushd']
 
 @contextmanager
 def pushd(dir):
-    ''' A context manager for pushing into a dir and automatically 
-    coming back to the previous one '''
+    '''A context manager (Python 2.5+ only) for stepping into a 
+    directory and automatically coming back to the previous one. 
+    The original directory is returned. Usage is like this::
+    
+        from __future__ import with_statement
+        # the above line is only needed for Python 2.5
+        
+        from paver.easy import *
+        
+        @task
+        def my_task():
+            with pushd('new/directory') as old_dir:
+                ...do stuff...
+    '''
     old_dir = os.getcwd()
     tasks.environment.info('cd %s' % dir)
     os.chdir(dir)